I have made quite a number of changes here.  Firstly, in the QuestionAnswerBank.h file, i have made changes to the type of value the variable index represents. Here it is set to an NSInteger and this was because of some errors that were present in my code which was related to the conflicting types of variables. I still receive some errors related to this, however i am currently dealing with
Moreover, I have removed the actions for the 4 button in the outlet collections and from the ThirdViewController and replaced it with one button (choicepressed) that carries out all 4 actions of the buttons. Each button in the thirdviewcontroller is assigned to the choicepressed actions. Instead, i am using the properties of the button to preform different functions. Each button has its own independent tag and this is set in the attributes section of the Third View Controller. Each of the button is given a tag from 1 to 4 and the same action will refer to all the 4 tags i.e. all the 4 buttons.

So in thirdviewcontroller.m, where sender.tag == 1, it is referring to the first button A and the same applies for sender.tag == 2.

In thirdviewcontroller.h, different variables have been added. There is now a variable for chosenanswer, actualanswer and questionnumber.  I have also added 2 other functions checkanswerfunc and next question. Both of these functions are defined in the thirdviewcontroller.m file and these functions are called under the button action.

Thus in ThirdViewController.m, when the sender.tag is a certain value, the chosen answer will be the buttons (in this case its specific tag) corresponding text (firstchoice.text, etc). Then the checkanswerfunc is carried out and this compares the chosen answer with the final answer and displays a notification label

At the moment, the code looks very disorganised however it is still work in progress and i hope to get the app working for all questions related to Numerical Reasoning and thus, it will be easier to program for Riddles and General Knowledge.

+++ BrainAppQuiz/QuestionAnswerBank.h
@@ -30,9 +30,9 @@
 @property (nonatomic, strong) NSString *result;
 
 
--(NSString*) printquestion: (NSString*) testtypefunc :(int) index;
--(NSString*) printanswerchoices: (NSString*) testtypefunc :(int) index;
--(NSString*) checkanswer: (NSString*) testtypefunc :(NSString*) answer :(int) index;
+-(NSString*) printquestion: (NSString*) testtypefunc :(NSInteger) index;
+-(NSString*) printanswerchoices: (NSString*) testtypefunc :(NSInteger) index;
+-(NSString*) checkanswer: (NSString*) testtypefunc :(NSString*) answer :(NSInteger) index;
 -(NSString*) nextquestionfunc;
